refactor(kitchenpos): dto의 구현 방식을 record로 변경
