+++ src/main/java/kitchenpos/products/application/dto/ChangeProductPriceRequestDto.java
@@ -3,21 +3,7 @@ package kitchenpos.products.application.dto;
 import java.math.BigDecimal;
 import kitchenpos.products.tobe.domain.model.ProductPrice;
 
-public class ChangeProductPriceRequestDto {
-
-    private BigDecimal price;
-
-    public ChangeProductPriceRequestDto() {
-
-    }
-
-    public ChangeProductPriceRequestDto(BigDecimal price) {
-        this.price = price;
-    }
-
-    public BigDecimal getPrice() {
-        return this.price;
-    }
+public record ChangeProductPriceRequestDto(BigDecimal price) {
 
     public ProductPrice toValueObject() {
         return new ProductPrice(price);
