+++ src/main/java/kitchenpos/products/application/dto/ChangeProductPriceResponseDto.java
@@ -4,37 +4,7 @@ import java.math.BigDecimal;
 import java.util.UUID;
 import kitchenpos.products.tobe.domain.model.Product;
 
-public class ChangeProductPriceResponseDto {
-
-    private UUID id;
-    private String name;
-    private BigDecimal price;
-
-    protected ChangeProductPriceResponseDto() {
-
-    }
-
-    public ChangeProductPriceResponseDto(
-        UUID id,
-        String name,
-        BigDecimal price
-    ) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-    }
-
-    public UUID getId() {
-        return this.id;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public BigDecimal getPrice() {
-        return this.price;
-    }
+public record ChangeProductPriceResponseDto(UUID id, String name, BigDecimal price) {
 
     public static ChangeProductPriceResponseDto from(Product product) {
         return new ChangeProductPriceResponseDto(
