+++ src/main/java/kitchenpos/products/application/dto/CreateProductRequestDto.java
@@ -2,28 +2,6 @@ package kitchenpos.products.application.dto;
 
 import java.math.BigDecimal;
 
-public class CreateProductRequestDto {
+public record CreateProductRequestDto(String name, BigDecimal price) {
 
-    private String name;
-    private BigDecimal price;
-
-    public CreateProductRequestDto() {
-
-    }
-
-    public CreateProductRequestDto(
-        String name,
-        BigDecimal price
-    ) {
-        this.name = name;
-        this.price = price;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    public BigDecimal getPrice() {
-        return this.price;
-    }
 }
